requirements.txt
beautifulsoup4
grass-gis-helpers
lxml
xmltodict
